requirements.txt
streamlit>=1.28.0
polars>=0.19.0
streamlit-molstar>=0.3.0
rdkit>=2023.3.1
luna
biopython>=1.81
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
plotly>=5.14.0